Mise à jour du projet React.js

+++ src/components/welcome.js
@@ -1,35 +1,53 @@
 import React, { useEffect, useState } from 'react';
-import { Link  } from 'react-router-dom';
+import {  useNavigate  } from 'react-router-dom';
 import Spinner from 'react-bootstrap/Spinner';
 
 
-const AppLauncher = () => {
+const Welcome = () => {
 
     const [isLoading, setIsLoading] = useState(true);
-	const [loading, setLoading] = useState(false);
+	const [appInfo, setAppInfo]     = useState(null);
+	const navigate = useNavigate();
 
-	const logOutAll = () => {
-		if (window.Android) {
-			setLoading(true);
-			window.Android.LogOutAll(() => {
-			  localStorage.clear();
-			  setLoading(false);
-			});
-		} else {
-			console.error('Android interface not found.');
-		}
-	};
-	
-	useEffect(() => {
+
+	const logout = () => {
+		setIsLoading(true);
 		setTimeout(() => {
-			setIsLoading(false);
-		}, 500); 
-    }, []);
+			if (window.Android && typeof window.Android.o00011 === 'function') {
+				window.Android.o00011(() => {
+				  localStorage.clear();
+				});
+			} else {
+				navigate('/un');
+			}
+		}, 1000);
+	};
+
+	useEffect(() => {
+        // Récupérer les données depuis localStorage dès que le composant est monté
+        const data = localStorage.getItem('AppInfo');
+        setTimeout(() => {
+			if (data) {
+				// Parsez les données JSON récupérées
+				const parsedData = JSON.parse(data);
+				
+				// Vérifiez la valeur de 'logged' et mettez à jour l'état
+				if (parsedData.logged === true) {
+					navigate('/welcome');
+				} else {
+					setAppInfo(parsedData);
+				}
+				setIsLoading(false);  // Mettre à jour le statut de chargement
+			} else {
+				setIsLoading(false);  // Pas de données, terminer le chargement
+			}
+		}, 1000); 
+    }, [appInfo, navigate]);
 	
 	// Pendant le chargement, vous pouvez afficher un indicateur visuel de chargement ou simplement retourner null
 	if (isLoading) {
-      return <div><><div className="loader fs-3 fw-normal text-dark p-5">
-		  <Spinner animation="grow" variant="warning" role="status">
+      return <div><><div className="loader fs-4 fw-normal text-dark p-5">
+		  <Spinner animation="border" variant="success" role="status">
 			  <span className="visually-hidden text-light">Loading...</span>
 		  </Spinner>
       </div></> </div>;
@@ -37,26 +55,38 @@ const AppLauncher = () => {
 
     return (
         <div className='App'>
-		    <nav className="navbar fixed-top">
-				<div className='container'>
-					<div className="navbar-left">
-						<Link to="/">
-						  <img src={`${process.env.PUBLIC_URL}/logo.png`} className="" alt="logo" /> 
-						</Link>
+     
+		    {appInfo ? (
+				<>
+					<nav className="navbar fixed-top pt-3">
+						<div className='container'>
+							<div className="d-flex">
+								<h4 className='fw-bold' >{appInfo.name}</h4>
+							</div>
+							<div className="d-flex float-right">
+								<button onClick={logout} className='logout'>
+								    <i className="fa-solid fa-power-off"></i>
+								</button>
+							</div>
+						</div>
+					</nav>
+					<div className='App-Container-Home'>
+						<div className="text-center">
+							<span className="text-dark fs-1 fw-bold"> Hello Word !</span> 
+						</div>
 					</div>
-					<div className="navbar-right">
-						<span className='fs-6 text-dark fw-bold'> Nouveau projet </span>
-					</div>
-				</div>
-			</nav>
-		    <div className='App-Container'>
-				<h1> hello Word !</h1> 
-				<button className='logout' onClick={logOutAll} >
-					{loading ? 'Chargement...' : 'Se déconnecter'}
-				</button>
-			</div>
+			    </>
+            ) : (
+			    <>
+				    <div className='App-Container'>
+						<div className="text-center">
+							<h2 className="text-danger">Erreur: Informations de l'application non disponibles</h2>
+						</div>
+			        </div>
+				</>
+            )}	   
         </div>     
     )
 };
 
-export default AppLauncher;
+export default Welcome;
